refactor. update pages construct
